UI: Hand Cursors
We are currently a bit confused regarding "hand" and "move" cursors. We
have a single define for WM_CURSOR_HAND which maps to a single ghost
cursor called GHOST_kStandardCursorMove. It is not correct to use a
"hand" to indicate move on Windows, and Mac needs a closed hand and a
pointing hand. This PR makes it so that WM_CURSOR_MOVE uses
GHOST_kStandardCursorMove, WM_CURSOR_HAND uses
GHOST_kStandardCursorHandOpen, WM_CURSOR_HAND_CLOSED uses
GHOST_kStandardCursorHandClosed, and WM_CURSOR_HAND_POINT uses
GHOST_kStandardCursorHandPoint. Although we don't use hand cursors on
Windows this adds cur files for these for completeness and testing.