Fix #127192: VSE Tonemap now works as expected on strips that don't cover whole screen
When doing average/min/max luminance calculations for tonemapping,
ignore pixels that are below 50% transparency. Due to how VSE innards
work, when a strip is positioned to not cover the whole screen,
"the rest" is filled with transparent black. For tonemapping, this
was dragging average calculated luminance way down.

Conceptually, just because something is not visible, does not make
the whole image "more dark"; it should be ignored instead.