Vulkan: Cannot use inside Renderdoc
Renderdoc only export extensions it supports.
`VK_EXT_dynamic_rendering_unused_attachments` is not supported by
renderdoc and therefor no devices could be found to start the vulkan
backend.

In GHOST we already work around this issue by not checking this specific
extension. We should do the same in VKBackend.