Color management: skip no-op colorspace transforms for float images (mostly affects VSE)
In VSE, the default setting is that even float images are all in sRGB
(or rather, display) space. However code that was drawing the final
VSE float image on screen, was first converting it to linear space,
and then back to display space.

An optimization already existed that skipped "no-op" conversions for
byte images. This adds a similar case for float images.

Note: I found a previous fix to an old issue (#39953, commit fe29f92030)
in related code. Probably should be double checked whether it has not
regressed.

VSE at 4K resolution, with a single 4K resolution EXR image strip,
playback on Ryzen5950X (Win10/VS2022):
- Playback average FPS 11 -> 18
- Time for sequencer_preview_region_draw part: 92ms -> 58ms
  - Time for sequencer_draw_display_buffer in there: 52ms -> 15ms