Vulkan: Add support for Intel Iris+ iGPUs
Intel Iris GPUs only support driver version 30/31 on Windows. These
drivers do not support VK_EXT_dynamic_rendering_unused_attachments.

The story is that this extension was added as validation layers was
complaining that it was needed. However renderdoc doesn't support
this extension and it was changed to be an optional extension. During
this last change the device detection was not implemented correctly
making the device detection only pass when the driver had support for
the extension but didn't check the device itself.

This is fixed by making the extension fully optional.