Vulkan: Incorrect Edge&Vertex selection
Edge and vertex selection selected the incorrect index. The reason is
that the extent of the area downloaded to the CPU for evaluation was
incorrect and read pixels could not in the place where they should
be stored in the CPU buffer.

Was an oversight when fixing face selection as here the pixels cover a
larger area the issue was less noticable.