Vulkan: Blendfile icon generation
Vulkan would crash when generating a file icon in case there is no
3d viewport in the active workspace. In this case the front buffer
color attachment was read. We didn't create a front buffer as swap
chains are handled differently.

We work around this issue in the same way as Metal does. Create a dummy
front framebuffer and share the same surface texture. When the icon
rendering reads from the front buffer it will read the data that was
created by the back buffer.