Fix #127399: Sculpt temporary Smooth brush ignores some settings
If the Smooth brush (which is switched to from other brushes holding
`Shift`) had the following settings (possibly others as well)
- Stabilize Stroke
- Airbrush
- Line
These were ignored (not drawing their correct cursor, airbrush not
working because the timer wasnt started).

Problem here is that `smooth_brush_toggle_on` (called via `stroke-
>test_start`) changes the brush, this was ignored in following code (and
in case of the `Stabilize Stroke` is called only after the code that was
checking if this was ON).

So to resolve, get the brush again after `stroke->test_start` has run
(and also move the check `Stabilize Stroke` a bit (to  place that
actually fits much better since similar things like line visualization
are handled there as well)