ShaderC: Remove globale locale lock
This is a backport of a patch in SPIRV-Tools that will allow faster
parallel compilation.

See https://github.com/KhronosGroup/SPIRV-Tools/issues/5802 for the
original report.

The issue is that SPIRV tools used a stringstream to convert an uint32_t
to a string hotloops when optimizing shaders. stdlib provided by the
platform would set a locale and would mutex this making parallel
compilation as fast as single threaded compilation.

This patch has been tested to build on Linux, but I haven't tested it on
Windows and macOs.