Refactor: Sculpt: Add face set offset parameter for BMesh functions
Part of 118145.

Even though BMesh face set functionality does not currently work, there
are some reasonable assumptions we can make about the needed parameters
for some APIs.

This commit changes the signature and introduces new methods that are
effecitvely stubs so that future changes are easier when / if face sets
are fully supported within BMesh.