Anim: add Action Slot selector to Action constraint
Add slotted Actions support to Action constraints.

The user interface still needs a bit of tweakage, as the properties
don't quite line up yet.

The unit test `bl_rigging_symmetrize` is also updated, to account for
the new properties on the Action constraint. It was doing floating point
comparisons on a collection property...