Nodes: Improve matches when dragging a node onto a link
This patch adds a new 'get_main_socket_by_type' function to match
the node sockets more precisely with the link.

This resolves some cases when nodes are unabel to drag link when there
are suitable sockets available. E.g. a menu socket in a group node would
prevent. This was previously investigated in #127717.

Previously, the `get_main_socket` function matched sockets as follows.
1) Socket declaration `is_default_link_socket` flag
2) Sockets matched by highest priority node socket*
3) Validate link

The new function matches sockets as follows.
1) Socket declaration `is_default_link_socket` flag (same as before)
2) Sockets matched by exact type with link socket*
3) Sockets matched by first lower type than link socket*
4) Sockets matched by highest priority node socket* (same as before)
5) Validate link (same as before)

* Visible node sockets, then hidden node sockets

Matching be name was not considered as it would have required
some hard coding of socket names.