Nodes: Expose is_default_link_socket option in group node
This exposes the `is_default_link_socket` socket declaration option
to the node group interface.

This allows a socket to be marked as the default link socket.

For example, a group node may have a Menu socket. This menu socket
prevents the ability to drag the group node over a Float link
when the main functionality may be for the Float.

If a group node has more than one socket marked with this option
then the socket higher in the list with same property take precedence.