Geometry Nodes: Add Seed value to Object Info node
This patch adds a seed output to the Object Info node which can
be used to drive randomness when duplicating objects with the
same node tree.

It does this by adding a persistent integer `Seed` value per object.
This seed is created when adding and also when duplicating objects.
The seed is generated as an uuid which is then hashed.
The object's seed value can be edited in the Object property panel.

This avoids issues with other methods such as basing randomness
on an objects location or manually editing a modifier input.