UI: Allow label for search template (respecting property split layout)
Since 5bcb0c9935, `uiLayout.template_ID()` already supports passing a
label, in which case the `uiLayout.use_property_split` and
`uiLayout.use_property_decorate` options would be respected. This
makes the template fit nicely with the surrounding layout, where buttons
might have a label split on the left, and a decorator button on the
right.

This change brings the same to `uiLayout.template_search()`.