UI: Win32 - Clipboard Image Page from File Path
On the Windows platform allow pasting an image into the Image Editor
based on a filepath in the clipboard. For example, selecting "copy" on
the context menu in the file system. This checks for CF_HDROP format
and should work for all image formats we support. This is being
considered because this type of workflow is very common on Mac and it
would be nice to match behavior on Windows.