Sculpt: Add Mask from Boundary Operator
This commit adds the `SCULPT_OT_MASK_FROM_BOUNDARY` operator, which
behaves similarly to `SCULPT_OT_MASK_FROM_CAVITY`. It shares much of the
same code and settings, and is conceptually similar: it takes the
automasking settings from either the Brush, the Scene, or defined by
the operator to apply the automask values to the currently existing mask
or creating one if one does not exist.