Fix #126521: Redo popup doesn't refresh layout on changes
98c92b9033 disabled refreshing for the redo popup, since that requires
passing an operator pointer to the UI to holding on to it while the popup is
visible. Usually operators are short lived and shouldn't be held by the UI, to
avoid dangling pointer accesses. In this case it's fine though, because the
operator will be kept alive in the window manager.

Partially reverts 98c92b9033, and adds a comment to note this special case.