UI Test: CJK TTC Face Selection
A test that could turn into a feature. Once applied this allows users
to select a font collection (TTC, OTC) file as the interface font. It
also allow selecting the face within the collection. So in the case
of Chinese, Japanese, Korean font collections this allows selecting a
specific language variant.