Anim: define RNA subtypes of ActionStrip, and refine to those types
This is a follow up to #126559. In that PR I temporarily removed the
strip subtypes from RNA because I didn't have the brain space to tackle
updating that at the same time.

This PR restores the strip subtypes in RNA, even though the strip
themselves don't exist as separate C++ types anymore. Now the RNA type
that we present the strip as depends on the type of strip data the strip
references.

Note that currently there is only one subtype: the keyframe strip.