UI: Use WM_CURSOR_EDIT For Corner Hover Cursor with Docking
Prior behavior when hovering over the corner action zones the mouse
cursor would change to WM_CURSOR_EDIT (like a plus sign). With docking
that is WM_CURSOR_HAND, but on other platforms it is WM_CURSOR_MOVE.
Unfortunately on some platforms WM_CURSOR_MOVE does not show a
different cursor from default so there isn't an change when hovering
the corner.  This PR changes this back to WM_CURSOR_EDIT for non-Mac
platforms.