Fix #127825: File Not Found: Nicer Alert & Do Not Clear Recent
When attempting to load a blend file that is not found or otherwise
invalid we are currently showing an (ugly) popup menu report and then
clearing the item from the Recent Files list. Clearing the list could
be unwanted in the case of files that are temporarily offline. And we
now have the ability to only clear "Items not found" using the "Clear
Recent Files List operator". This PR removes that clearing and shows
a nicer alert dialog (if not running in background).