Anim: update some uses of listbases of fcurves
Because we're moving to layered actions, which don't store their
fcurves in a list base, we need to update the places that assume the old
listbase-based structure.

This commit addressed the low-hanging fruit where it was fairly obvious
how to correctly update the code with minimal changes. Other cases that
either weren't immediately obvious or required non-trivial code changes
(or both) have been left for future PRs.