Vulkan: Add support for VK_KHR_fragment_shader_barycentric
This commit introduces support for the
`VK_KHR_fragment_shader_barycentric` extension. If the physical device
supports the extension, it is enabled and used directly. Otherwise, a
workaround is applied by injecting a geometry shader to generate
barycentric coordicates, which was the previous behavior.

Besides, `gpu_pos` and `gpu_pos_flat` are not needed if using this
extension, since the order of the vertices is defined in the spec,
whereas `GL_AMD_shader_explicit_vertex_paramete` would still need `gpu_pos`
to be passed around for stable barycentric coordinate.