Vulkan: Fix location mismatch by separating input and output locations
If we don't separate input and output locations, then the location count
of workarounds will mismatch between stages, if multiple workarounds are
enabled at the same time.

For example, if do_layer_workaround and do_barycentric_workaround are
enabled  at the same time, the `location` of `gpu_BaryCoord` in fragment
shader will mismatch with that in the inject geometry shader.