Option to keep world transform when pasting from clipboard.
This is an option for the paste operator. Clone tool could use a similar
option via tool settings, if necessary, although cloning usually happens
within the same object (?) so object-object transform doesn't make a
difference.

The object-world transform is stored separately when copying to the
clipboard and not applied to avoid unnecessary loss of accuracy.
The transform is applied when copying to a different object and the
world transform of the copied strokes should remain unchanged.