VSE: Prototype: Remove timeline box select
This patch removes the box select tool from the timeline to simplify
selection logic, avoid confusion for users, and make maintaining the
code easier in the future.

Since both the box select and tweak tools had nearly identical behavior,
with both being able to tweak, individually select, and box select,
removing one in favor of the other should standardize behavior and
reduce the likelihood of bugs appearing (e.g. d2091b4b1 ended up
breaking connected strips because it tried to copy functionality already
present in `sequencer_select_exec()`).

Simplifications include:
- `tweak` property is no longer needed for `sequencer.select_box` to
  select a strip before a `transform.seq_slide` occurs
- Redundant `sequencer.select` click item in Tweak tool removed. This
  had `select_passthrough` disabled, but `select_passthrough` isn't
  actually used anywhere in sequencer code, so it has no meaning
- Redundant `transform.translate` item in Tweak fallback removed, which
  was preventing a box-select from happening if it was the fallback
  tool. This will make it possible for non-selection tools to still be
  able to box-select if drag is not already handled
  - Note that before this change, the last used tool between box select
    and tweak would effectively define the fallback tool. In the 3D
    viewport this can be explicitly selected with the "Drag" tool
    setting in the header, but since that's not present here, users
    would have had restrictions when trying to use e.g. blade or other
    future tools (either tweak or box select but not both).

There should be no regressions, but some bugs have been fixed with these
changes. Preview box select remains as-is.

To consider:
In RCS, the box select tool has a small unique feature, in
that it allows the user to select with left-click-drag instead of
right-click-drag.

With the tool gone, this behavior is no longer possible, however, the
keymap is called "right-click select" so it makes sense to me that a box
select could be triggered by right-click-drag w/ the tweak tool/fallback
tweak tool.